4.14 blogilistan laajennus, step2

+++ controllers/blogs.js
@@ -19,8 +19,34 @@ blogsRouter.post('/', async (req, res, next) => {
 
 blogsRouter.delete('/:id', async (req, res, next) => {
   try {
-    await Blog.findByIdAndDelete(req.params.id)
-    res.status(204).end()
+    const blogExists = await Blog.findById(req.params.id)
+    if (blogExists !== null) {
+      await Blog.findByIdAndDelete(req.params.id)
+      res.status(204).end()
+    } else {
+      res.status(400).end()
+    }
+  } catch (e) {
+    next(e)
+  }
+})
+
+blogsRouter.put('/:id', async (req, res, next) => {
+  const editedBlog = {
+    title: req.body.title,
+    author: req.body.author,
+    url: req.body.url,
+    likes: req.body.likes
+  }
+
+  try {
+    const blogExists = await Blog.findById(req.params.id)
+    if (blogExists !== null) {
+      const updatedBlog = await Blog.findByIdAndUpdate(req.params.id, editedBlog, { new: true })
+      res.status(200).json(updatedBlog)
+    } else {
+      res.status(400).end()
+    }
   } catch (e) {
     next(e)
   }
